feat(moon-display): Add moon display
Added a feature that displays a moon icon during the night in the current weather MainPanel. This feature is activated by adding the 'isDay' prop to the MainPanel component. In the WeeklyForecastModal, the 'isDay' prop is set to 1 (which ensures that only day icons are displayed).
Additionally: the TodaysWeather component has been removed as it became unnecessary after adding the MainPanel component.